% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_comb.R
\name{all_comb}
\alias{all_comb}
\title{Analyze all combinations}
\usage{
all_comb(gdist.inputs = NULL,
                jl.inputs = NULL,
                GA.inputs,
                results.dir,
                max.combination = 4,
                iters = 1000,
                replicate = 1,
                sample.prop = 0.75,
                nlm = FALSE,
                dist_mod = TRUE,
                null_mod = TRUE,
                ...)
}
\arguments{
\item{gdist.inputs}{Object created from running
\code{\link[ResistanceGA2]{gdist.prep}}. Supply when optimizing with
\pkg{gdistance}.}

\item{jl.inputs}{Object created from running
\code{\link[ResistanceGA2]{jl.prep}}. Supply when optimizing with
Circuitscape in Julia.}

\item{GA.inputs}{Object created from running
\code{\link[ResistanceGA2]{GA.prep}}. Be sure that
\code{Results.dir = "all.comb"} was used when creating this object.}

\item{results.dir}{Directory to write and save analysis results. If the
directory does not exist, it will be created automatically. Any existing
files located in this directory may be deleted after confirmation.}

\item{max.combination}{The maximum number of surfaces to include in the all combinations analysis (Default = 4). Alternatively, specify a vector with the minimum and maximum number of surfaces to combine (e.g., c(2,4). If the minimum > 1, then the single surface optimization will be skipped.}

\item{iters}{Number of bootstrap iterations to be conducted (Default = 1000)}

\item{replicate}{The number of times to replicate the GA optimization process for each surface (Default = 1)}

\item{sample.prop}{Proportion of observations to be sampled each iteration (Default = 0.75)}

\item{nlm}{(NOT CURRENTLY IMPLEMENTED) Logical, if TRUE, the final step of optimization will use nlm to fine-tune parameter estimates. This may lead to overfitting in some cases. (Default = FALSE)}

\item{dist_mod}{Logical, if TRUE, a Distance model will be calculated and added to the output table (default = TRUE)}

\item{null_mod}{Logical, if TRUE, an intercept-only model will be calculated and added to the output table (Default = TRUE)}

\item{...}{Additional arguments (Not currently used)}
}
\value{
An object of class \code{resga_all_comb}. This function optimizes
  resistance surfaces in isolation using
  \code{\link[ResistanceGA2]{SS_optim}}, followed by multisurface
  optimization using \code{\link[ResistanceGA2]{MS_optim}}, and then
  conducts a bootstrap analysis.
}
\description{
A wrapper function to run both \code{\link[ResistanceGA2]{SS_optim}} and
\code{\link[ResistanceGA2]{MS_optim}} to optimize all combinations of
resistance surfaces with the Genetic Algorithms package \pkg{GA}. Following
optimization, \code{\link[ResistanceGA2]{Resist.boot}} is run to conduct a
bootstrap analysis.
}
\examples{
\dontrun{
pts <- terra::vect(sample_pops[[1]], type = "points")
gd <- lower(Dc_list[[1]])

gdist.inputs <- gdist.prep(
  n.Pops = nrow(sample_pops[[1]]),
  response = gd,
  samples = pts
)

GA.inputs <- GA.prep(
  raster = raster_orig,
  Results.dir = "all.comb",
  quiet = TRUE
)

results_dir <- file.path(tempdir(), "ResistanceGA2-all-comb")

combo.out <- all_comb(
  gdist.inputs = gdist.inputs,
  GA.inputs = GA.inputs,
  results.dir = results_dir,
  max.combination = 2,
  iters = 10,
  replicate = 1
)
}
}
\author{
Bill Peterman <Peterman.73@osu.edu>
}
