% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostic_Plots.R
\name{Diagnostic.Plots}
\alias{Diagnostic.Plots}
\title{Create diagnostic plots}
\usage{
Diagnostic.Plots(resistance.mat, genetic.dist, XLAB,YLAB, plot.dir, type, name, ID, ZZ)
}
\arguments{
\item{resistance.mat}{Path to a Circuitscape resistance output file, or a
matrix/vector of pairwise resistance distances such as the output from
\code{\link{Run_gdistance}}.}

\item{genetic.dist}{Vector of pairwise genetic distances (lower half of pairwise matrix). Can be accessed from \code{gdist.inputs$response} or \code{jl.inputs$response}}

\item{XLAB}{Label for x-axis (Defaults to "Estimated resistance")}

\item{YLAB}{Label for y-axis (Defaults to "Genetic distance")}

\item{plot.dir}{Directory to output TIFF of diagnostic plots}

\item{type}{Specify whether the optimized surface is "continuous" or "categorical"}

\item{name}{The name to be attached to the output file. Must be specified when getting diagnostic plots for gdistance models}

\item{ID}{The to_from ID list for the MLPE model. The function will automatically create this object, but it can be specified directly from the output of \code{gdist.prep} or \code{jl.prep} (Default = NULL)}

\item{ZZ}{The sparse matrix object for the MLPE model. The function will automatically create this object, but it can be specified directly from the output of \code{gdist.prep} or \code{jl.prep} (Default = NULL)}
}
\value{
A multipanel panel .tif including histogram of residuals and qqplot of fitted mixed effects model
}
\description{
This function will generate mixed effect model diagnostic plots following optimization
}
\examples{
\dontrun{
Diagnostic.Plots(
  resistance.mat = resist_list[[1]],
  genetic.dist = lower(Dc_list[[1]]),
  plot.dir = paste0(normalizePath(tempdir(), winslash = "/"), "/"),
  type = "categorical",
  name = "example"
)
}
}
\author{
Bill Peterman <Peterman.73@osu.edu>
}
