% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{jl_out}
\alias{jl_out}
\title{Data file. Example raster output from CIRCUITSCAPE Julia}
\format{
\code{SpatRaster} with 1 layer
}
\usage{
data(jl_out)
}
\description{
A \code{SpatRaster} produced by a Circuitscape Julia run.
}
\keyword{datasets}
