% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLPE.R
\name{MLPE.lmm}
\alias{MLPE.lmm}
\title{Run maximum likelihood population effects mixed effects model (MLPE)}
\usage{
MLPE.lmm(resistance, 
                pairwise.genetic, 
                REML = FALSE, 
                ID = NULL, 
                ZZ = NULL,
                scale = TRUE)
}
\arguments{
\item{resistance}{Path to a Circuitscape resistance output file, or a
vector, matrix, or \code{dist} object of pairwise resistance distances.}

\item{pairwise.genetic}{Lower half of pairwise genetic distance matrix}

\item{REML}{Logical. If TRUE, mixed effects model will be fit using restricted maximum likelihood. Default = FALSE}

\item{ID}{The to_from ID list for the MLPE model. The function will automatically create this object, but it can be specified directly from the output of \code{gdist.prep} or \code{jl.prep} (Default = NULL)}

\item{ZZ}{The sparse matrix object for the MLPE model. The function will automatically create this object, but it can be specified directly from the output of \code{gdist.prep} or \code{jl.prep} (Default = NULL)}

\item{scale}{Specify whether the pairwise distance values be scaled and centered (Default = TRUE)}
}
\value{
A lmer object from the fitted model
}
\description{
Runs MLPE as detailed by Clarke et al. (2002). This function will run the model and return lmer object
}
\details{
An AIC value will only be returned if \code{REML = FALSE}
}
\examples{
\dontrun{
fit <- MLPE.lmm(
  resistance = lower(resist_list[[1]]),
  pairwise.genetic = lower(Dc_list[[1]])
)

summary(fit)
}
}
\references{
Clarke, R. T., P. Rothery, and A. F. Raybould. 2002. Confidence limits for regression relationships between distance matrices: Estimating gene flow with distance. Journal of Agricultural, Biological, and Environmental Statistics 7:361-372.
}
\author{
Bill Peterman <Peterman.73@osu.edu>
}
