% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_optim.R
\name{MS_optim}
\alias{MS_optim}
\title{Simultaneous optimization of multiple resistance surfaces}
\usage{
MS_optim(gdist.inputs = NULL, jl.inputs = NULL, GA.inputs, diagnostic_plots = TRUE)
}
\arguments{
\item{gdist.inputs}{Object from \code{\link{gdist.prep}}. Supply when optimizing with gdistance.}

\item{jl.inputs}{Object from \code{\link{jl.prep}}. Supply when optimizing with Circuitscape via Julia.}

\item{GA.inputs}{Object from \code{\link{GA.prep}}.}

\item{diagnostic_plots}{Logical. Generate and save diagnostic plots? Default = \code{TRUE}.}
}
\value{
A named list with GA summary, MLPE model, AICc table, distance matrices,
  percent contribution, and \code{k} table.
}
\description{
Optimize multiple resistance surfsaces simultaneously using genetic algorithms
}
\examples{
 
## Not run:
## *** TO BE COMPLETED *** ##

## End (Not run)
}
\author{
Bill Peterman <Peterman.73@osu.edu>
}
