% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{raster_orig}
\alias{raster_orig}
\title{Data file. Simulated original resistance surfaces}
\format{
\code{SpatRaster} with 2 layers
}
\usage{
data(raster_orig)
}
\description{
A \code{SpatRaster} with two layers generated from the \code{RandomFields}
package. These are the raw (untransformed) surfaces used as optimization inputs.
}
\details{
\itemize{
  \item cat_orig. 5-class categorical resistance surface.
  \item cont_orig. Continuous resistance surface.
}
}
\keyword{datasets}
