% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{raster_true}
\alias{raster_true}
\title{Data file. Simulated true resistance surfaces}
\format{
\code{SpatRaster} with 3 layers
}
\usage{
data(raster_true)
}
\description{
A \code{SpatRaster} with three layers generated from the \code{RandomFields}
package. These surfaces were used with individual-based genetic simulations
conducted using \code{PopGenReport}.
}
\details{
\itemize{
  \item cat_true. 4-class categorical resistance surface. Reclass: 1=1, 2=75, 3=200, 4=25, 5=1.
  \item cont_true. Continuous resistance surface transformed via inverse monomolecular
    (shape = 2.75, max = 100).
  \item multi_true. Composite surface: sum of transformed categorical and continuous
    surfaces, rescaled to minimum of 1.
}
}
\keyword{datasets}
