% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{resistance_surfaces}
\alias{resistance_surfaces}
\title{Data file. Simulated resistance surfaces}
\format{
\code{SpatRaster} with 3 layers
}
\usage{
data(resistance_surfaces)
}
\description{
A \code{SpatRaster} with three layers for use in vignette examples.
}
\details{
\itemize{
  \item categorical. 3-class categorical resistance surface.
  \item continuous. Continuous resistance surface.
  \item feature. 2-class categorical (feature) resistance surface.
}
}
\keyword{datasets}
