% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{package}
\name{ResistanceGA2-package}
\alias{ResistanceGA2}
\alias{ResistanceGA2-package}
\title{ResistanceGA2: Optimize Resistance Surfaces Using Genetic Algorithms}
\description{
Provides functions to optimize resistance surfaces using Genetic Algorithms. Continuous and categorical surfaces can be optimized, and multiple surfaces can be simultaneously optimized to create novel resistance surfaces. Resistance distances can be calculated via Circuitscape run in Julia or via the gdistance package.
}
\details{
This package provides functions to optimize continuous and categorical
resistance surfaces using
\href{https://github.com/Circuitscape/Circuitscape.jl}{Circuitscape.jl}
(run through Julia) and Genetic Algorithms within R.
Resistance distances can alternatively be computed using the
\pkg{gdistance} package (least-cost paths / commute distances).

Output includes: a summary table with AIC, AICc, conditional and marginal
R-squared values, and log-likelihood values for each optimized surface;
parameters that optimized each top model; coefficients from fitted mixed
effects models; response-curve plots; and diagnostic plots of model fit.
}
\references{
Peterman, W.E., G.M. Connette, R.D. Semlitsch, and L.S. Eggert. 2014.
Ecological resistance surfaces predict fine-scale genetic differentiation in
a terrestrial woodland salamander. Molecular Ecology 23:2402--2413.
\doi{10.1111/mec.12708}

Peterman, W. E. 2018. ResistanceGA: An R package for the optimization of
resistance surfaces using genetic algorithms. Methods in Ecology and
Evolution 9:1638--1647. \doi{10.1111/2041-210X.12984}
}
\author{
Bill Peterman \email{peterman.73@osu.edu}
}
\keyword{internal}
