% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{rga2_demo_covariates}
\alias{rga2_demo_covariates}
\title{Data file. Idaho demo raster covariates}
\format{
\code{SpatRaster} with 6 layers
}
\usage{
data(rga2_demo_covariates)
}
\description{
A \code{SpatRaster} containing the six Idaho raster layers used to construct
\code{rga2_demo}. The layers are derived from \pkg{GeNetIt} and include
\code{cti}, \code{err27}, \code{ffp}, \code{gsp}, \code{hli}, and
\code{nlcd}.
}
\keyword{datasets}
