% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{rga2_demo_genetic}
\alias{rga2_demo_genetic}
\title{Data file. Idaho demo genetic distance matrix}
\format{
A 12 x 12 numeric matrix
}
\usage{
data(rga2_demo_genetic)
}
\description{
A 12 x 12 symmetric chord-distance matrix used in \code{rga2_demo}. Row and
column names match the \code{demo_id} values in \code{rga2_demo$sites}.
}
\keyword{datasets}
