% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_gdistance.R
\name{Run_gdistance}
\alias{Run_gdistance}
\title{Run gdistance cost/commute distance calculation}
\usage{
Run_gdistance(gdist.inputs, r, scl = TRUE)
}
\arguments{
\item{gdist.inputs}{Object created by \code{\link[ResistanceGA2]{gdist.prep}}.}

\item{r}{A \code{SpatRaster} object (from the \pkg{terra} package) or the
path to a raster file readable by \code{terra::rast()}.}

\item{scl}{Logical. Scale the correction values (default \code{TRUE}). Set
to \code{FALSE} to obtain absolute distance values. See
\code{\link[gdistance]{geoCorrection}} for details.}
}
\value{
A numeric distance vector (or matrix) of pairwise cost/commute
  distances. Returns \code{-99999} on error or warning.
}
\description{
Calculates pairwise cost or commute distances across a resistance surface
using the \code{gdistance} package.
}
\examples{
\dontrun{
library(terra)
r <- rast(nrows = 50, ncols = 50, vals = runif(2500))
coords <- matrix(runif(20, 0, 50), ncol = 2)
gdist.inputs <- gdist.prep(n.Pops = 10, samples = coords)
cd <- Run_gdistance(gdist.inputs, r)
}
}
\author{
Bill Peterman <Peterman.73@osu.edu>
}
