% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{sample_pops}
\alias{sample_pops}
\title{Data file. A list of three coordinate matrices}
\format{
A list of length 3; each element is a 50 x 2 coordinate matrix.
}
\usage{
data(sample_pops)
}
\description{
50 sample locations per list element corresponding to populations used in
genetic simulations. Each element is a two-column numeric matrix (x, y).
}
\details{
\itemize{
  \item sample_cat. Sample locations for categorical surface simulation.
  \item sample_cont. Sample locations for continuous surface simulation.
  \item sample_multi. Sample locations for multi-surface simulation.
}
}
\keyword{datasets}
