% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_optim.R
\name{SS_optim}
\alias{SS_optim}
\title{Single surface optimization}
\usage{
SS_optim(gdist.inputs,
 jl.inputs,
 GA.inputs,
 dist_mod, 
 null_mod,
 diagnostic_plots = TRUE)
}
\arguments{
\item{gdist.inputs}{Object created from running \code{\link[ResistanceGA2]{gdist.prep}} function. Defined if optimizing using gdistance}

\item{jl.inputs}{Object created from running \code{\link[ResistanceGA2]{jl.prep}} function. Defined if optimizing using CIRCUITSCAPE run in Julia}

\item{GA.inputs}{Object created from running \code{\link[ResistanceGA2]{GA.prep}} function}

\item{dist_mod}{Logical, if TRUE, a Distance model will be calculated and added to the output table (default = TRUE)}

\item{null_mod}{Logical, if TRUE, an intercept-only model will be calculated and added to the output table (default = TRUE)}

\item{diagnostic_plots}{Plotting and saving of diagnostic plots (Default = TRUE)}
}
\value{
This function optimizes resistance surfaces in isolation. Following optimization of all surfaces, several summary objects are created.\cr
\enumerate{
\item Diagnostic plots of model fit are output to the \code{Results/Plots}
directory created beneath \code{Results.dir}.
\item A .csv file with the Maximum Likelihood Population Effects mixed effects model coefficient estimates (MLPE_coeff_Table.csv)
\item Three summary .csv files are generated: CategoricalResults.csv, ContinuousResults.csv, & All_Results_AICc.csv. These tables contain AICc values and optimization summaries for each surface.
}
All results tables are also summarized in a named list ($ContinuousResults, $CategoricalResults, $AICc, $MLPE, $MLPE.list, $cd, $k)\cr
The \code{lmer} model objects stored $MLPE.list are fit using Restricted Maximum Likelihood \cr
$cd is a list of the optimized cost pairwise distance matrices and $k is a table of the surface names and number of parameters used to calculate AICc. These two objects can be passed to \code{\link[ResistanceGA2]{Resist.boot}} to conduct a bootstrap analysis.
}
\description{
Optimize each surface in a \code{SpatRaster} stack using a genetic algorithm
executed with the \code{\link[GA]{ga}} function in \pkg{GA}.
}
\examples{
\dontrun{
pts <- terra::vect(sample_pops[[1]], type = "points")
gdist.inputs <- gdist.prep(
  n.Pops = nrow(sample_pops[[1]]),
  response = lower(Dc_list[[1]]),
  samples = pts
)

GA.inputs <- GA.prep(
  raster = raster_orig,
  Results.dir = file.path(tempdir(), "ResistanceGA2-ss"),
  pop.size = 20,
  maxiter = 10,
  run = 5,
  quiet = TRUE,
  monitor = FALSE
)

ss.out <- SS_optim(
  gdist.inputs = gdist.inputs,
  GA.inputs = GA.worker.inputs,
  dist_mod = FALSE,
  null_mod = FALSE,
  diagnostic_plots = FALSE
)
}
}
\author{
Bill Peterman <Peterman.73@osu.edu>
}
