% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_optim.R
\name{MS_optim}
\alias{MS_optim}
\title{Simultaneous optimization of multiple resistance surfaces}
\usage{
MS_optim(gdist.inputs = NULL, jl.inputs = NULL, GA.inputs, diagnostic_plots = TRUE)
}
\arguments{
\item{gdist.inputs}{Object from \code{\link{gdist.prep}}. Supply when optimizing with gdistance.}

\item{jl.inputs}{Object from \code{\link{jl.prep}}. Supply when optimizing with Circuitscape via Julia.}

\item{GA.inputs}{Object from \code{\link{GA.prep}}.}

\item{diagnostic_plots}{Logical. Generate and save diagnostic plots? Default = \code{TRUE}.}
}
\value{
An object of class \code{resga_ms_optim}: a named list with GA
  summary, MLPE model, AICc table, distance matrices, percent contribution,
  and \code{k} table.
}
\description{
Optimize multiple resistance surfsaces simultaneously using genetic algorithms
}
\examples{
\dontrun{
pts <- terra::vect(sample_pops[[1]], type = "points")
gdist.inputs <- gdist.prep(
  n.Pops = nrow(sample_pops[[1]]),
  response = lower(Dc_list[[1]]),
  samples = pts
)

GA.inputs <- GA.prep(
  raster = raster_orig,
  Results.dir = file.path(tempdir(), "ResistanceGA2-ms"),
  pop.size = 20,
  maxiter = 10,
  run = 5,
  quiet = TRUE,
  monitor = FALSE
)

ms.out <- MS_optim(
  gdist.inputs = gdist.inputs,
  GA.inputs = GA.worker.inputs,
  diagnostic_plots = FALSE
)
}
}
\author{
Bill Peterman <Peterman.73@osu.edu>
}
