% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_from.R
\name{To.From.ID}
\alias{To.From.ID}
\title{Convenience function to make to-from object needed for specifying MLPE random effects}
\usage{
To.From.ID(sampled_pops,
                  pop_n = NULL,
                  spLoc = NULL,
                  nb = NULL,
                  ...)
}
\arguments{
\item{sampled_pops}{Integer. Number of populations (or individuals) sampled.}

\item{pop_n}{Integer vector of length \code{sampled_pops} giving the number
of individuals sampled at each population. Only needed when converting
from population- to individual-level analysis.}

\item{spLoc}{Sample locations, used to add a spatial-proximity random effect.
Accepts a two-column coordinate matrix (x, y), a
\code{\link[terra]{SpatVector}} of points, or a path to a tab-delimited
file (site, x, y) in Circuitscape format. Requires \code{nb}.}

\item{nb}{Maximum distance (in the same units as \code{spLoc}) below which
locations are considered to be in the same neighborhood. Must be specified
when \code{spLoc} is provided.}

\item{...}{Additional arguments. Currently only \code{dist} is accepted.}
}
\value{
A data frame of population pairs for use as the MLPE random effect.
}
\description{
Convenience function to make to-from object needed for specifying MLPE random effects
}
\details{
Distance values must come from the lower half of a distance matrix; use
\code{\link[ResistanceGA2]{lower}} to extract them. When fitting a MLPE
model with \code{\link[ResistanceGA2]{mlpe_rga}}, add \code{pop1} from this
object to the model data frame and specify it as the random effect.
}
\examples{
## Standard use
To.From.ID(sampled_pops = 4)

## Population to individual-level
To.From.ID(sampled_pops = 4, pop_n = c(1, 2, 3, 1))

## Expand pairwise distances
To.From.ID(sampled_pops = 4, pop_n = c(1, 2, 3, 1), dist = runif(6))

## Spatial-proximity random effect via coordinate matrix
spLoc <- cbind(c(1, 2, 9, 10), c(1, 2, 9, 10))
To.From.ID(sampled_pops = 4, spLoc = spLoc, nb = 3)

}
\author{
Bill Peterman <Peterman.73@osu.edu>
}
