% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceGA.R
\docType{data}
\name{vignette_current_map}
\alias{vignette_current_map}
\title{Data file. Precomputed Circuitscape current map used in vignette examples}
\format{
\code{SpatRaster} with 1 layer
}
\usage{
data(vignette_current_map)
}
\description{
A \code{SpatRaster} cumulative current map produced by
\code{\link{Run_CS.jl}} for the \code{"continuous"} layer of
\code{\link{resistance_surfaces}} (after \code{terra::unwrap()} when
needed) and \code{samples}.
}
\keyword{datasets}
